Refactor Categories table

+++ database/migrations/2024_05_31_095323_create_categories_table.php
@@ -28,6 +28,8 @@ return new class extends Migration
      */
     public function down()
     {
+        if(!app()->isLocal()) return;
+
         Schema::dropIfExists('categories');
     }
 };
